requirements.txt
alembic==1.7.5
aniso8601==9.0.1
black==21.12b0
boto3==1.20.26
botocore==1.23.26
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-API==3.0.post1
Flask-HTTPAuth==4.5.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-enum==1.5.1
mypy-extensions==0.4.3
password-strength==0.0.3.post2
pathspec==0.9.0
platformdirs==2.4.0
psycopg2==2.9.2
PyJWT==2.3.0
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
SQLAlchemy==1.4.28
tomli==1.2.3
typing_extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
